feat: allow updating and deleting group

+++ src/app/groups/[id]/scores.tsx
@@ -94,7 +94,7 @@ const Scores = () => {
               handleCancel={handleCancelAddScore}
               handleAdd={handleAddScore}
               onNameChange={setSubject}
-              onScoreChange={(score) => setScore(parseInt(score))}
+              onScoreChange={(score) => setScore(parseFloat(score))}
             />
           }
         </View>
